CDVT [fix][add] SucoThietbiController, add EquipWithDepart_Result, GetEquipmetIncident_Result
