+++ QUANGHANH2/EntityResult/EquipWithDepart_Result.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace QUANGHANH2.EntityResult
+{
+    public class EquipWithDepart_Result
+    {
+        public string equipment_id { get; set; }
+        public string equipment_name { get; set; }
+        public string department_name { get; set; }
+    }
+}
