+++ QUANGHANH2/EntityResult/GetEquipmetIncident_Result.cs
@@ -1,0 +1,47 @@
+﻿using QUANGHANH2.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace QUANGHANH2.EntityResult
+{
+    public class GetEquipmetIncident_Result : Incident1
+    {
+        public int? time_different { get; set; }
+        public string equipment_name { get; set; }
+        public string department_name { get; set; }
+        public string Equipment_category_id { get; set; }
+        public string mark_code { get; set; }
+        public string fabrication_number { get; set; }
+        public string stringStartTime { get; set; }
+        public string stringEndTime { get; set; }
+        public string stringDiffTime { get; set; }
+        public string editAble { get; set; }
+
+        public string getEndtime()
+        {
+            if (end_time == null) return "";
+            else
+            {
+                DateTime.TryParse(end_time.ToString(), out DateTime temp);
+                return temp.ToString("HH:mm dd/MM/yyyy");
+            }
+        }
+
+        public string getDiffTime()
+        {
+            if (end_time == null) return "";
+            else
+            {
+                DateTime.TryParse(end_time.ToString(), out DateTime temp);
+                TimeSpan timespan = temp.Subtract(start_time);
+                string output = "";
+                if (timespan.Days != 0) output += timespan.Days + " ngày ";
+                if (timespan.Hours != 0) output += timespan.Hours + " giờ ";
+                if (timespan.Minutes != 0) output += timespan.Minutes + " phút ";
+                return output;
+            }
+        }
+    }
+}
