+++ QUANGHANH2/QUANGHANH2.csproj
@@ -357,6 +357,8 @@
     <Compile Include="Controllers\TCLD\Transfer\TransferController.cs" />
     <Compile Include="Controllers\TCLD\Employees\WorkHistoryController.cs" />
     <Compile Include="Controllers\TCLD\Work\WorkGroupTypeController.cs" />
+    <Compile Include="EntityResult\EquipWithDepart_Result.cs" />
+    <Compile Include="EntityResult\GetEquipmetIncident_Result.cs" />
     <Compile Include="EntityResult\GetIncident_IncidentByDate_Result.cs" />
     <Compile Include="EntityResult\GetIncident_IncidentByYear_Result.cs" />
     <Compile Include="EntityResult\GetIncident_IncidentDB_Result.cs" />
